p3_util

Function iter_array_chunks_padded

pub fn iter_array_chunks_padded<T: Copy, const N: usize>(
    iter: impl IntoIterator<Item = T>,
    default: T,
) -> impl Iterator<Item = [T; N]>

Returns an iterator over N elements of the iterator at a time.

The chunks do not overlap. If N does not divide the length of the iterator, then the last N-1 elements will be padded with the given default value.

This is essentially a copy pasted version of the nightly array_chunks function. https://doc.rust-lang.org/std/iter/trait.Iterator.html#method.array_chunks Once that is stabilized this and the functions above it should be removed.