p3_util

Function reverse_slice_index_bits

pub fn reverse_slice_index_bits<F>(vals: &mut [F])
where
    F: Copy + Send + Sync,

Permutes arr such that each index is mapped to its reverse in binary.

If the whole array fits in fast cache, then the trivial algorithm is cache friendly. Also, if T is really big, then the trivial algorithm is cache-friendly, no matter the size of the array.